notas 10

+++ src/practica4/IPickable.kt
@@ -2,6 +2,6 @@ package practica4
 
 interface IPickable {
     val Name: String
-    var Weight: Int
+    var Weight: Double
     var Slots: Int
 }
